% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{amrrules_save}
\alias{amrrules_save}
\title{Save AMR Geno-Pheno Analysis Results and Plots to Files and Draft AMRrules}
\usage{
amrrules_save(
  amrrules,
  width = 9,
  height = 9,
  dir_path,
  outdir_name = NULL,
  file_prefix = NULL,
  minObs = 3,
  low_threshold = 20,
  bp_site = NULL,
  ruleID_start = 1000,
  mic_S = NULL,
  mic_R = NULL,
  disk_S = NULL,
  disk_R = NULL,
  use_mic = TRUE,
  use_disk = TRUE,
  guide = "EUCAST 2025",
  makeRules = TRUE,
  testRules = TRUE,
  expected_R = NULL,
  expected_I = NULL
)
}
\arguments{
\item{amrrules}{A list containing the output of the \code{\link{amrrules_analysis}} function, including plots and statistical summaries.}

\item{width}{Numeric, the width of the output plot images (default: 9).}

\item{height}{Numeric, the height of the output plot images (default: 9).}

\item{dir_path}{The path to the directory where the output will be saved.}

\item{outdir_name}{The name of the subdirectory within \code{dir_path} to save the results in (default: the antibiotic name).}

\item{file_prefix}{The prefix for filenames of output files (default: the antibiotic name).}

\item{minObs}{The minimum number of observations required to include a rule (default: 3).}

\item{low_threshold}{The threshold below which rules are assigned to evidence grade 'low' (default: 20).}

\item{bp_site}{The breakpoint site to filter on (Optional, use the \code{\link{summarise_data}} function to check whether there are different breakpoints for different sites and choose which one to specify here. By default, if multiple breakpoints are available the most conservative will be used.).}

\item{ruleID_start}{The starting numeric ID with which to number rules (integer, default: 1000).}

\item{mic_S}{The MIC breakpoint to define S (Optional, by default breakpoints are extracted using the AMR package, however if none are available or user wants to use a different one e.g. an ECOFF then it can be specified here).}

\item{mic_R}{The MIC breakpoint to define R (Optional, by default breakpoints are extracted using the AMR package, however if none are available or user wants to use a different one e.g. an ECOFF then it can be specified here).}

\item{disk_S}{The disk breakpoint to define S (Optional, by default breakpoints are extracted using the AMR package, however if none are available or user wants to use a different one e.g. an ECOFF then it can be specified here).}

\item{disk_R}{The disk breakpoint to define R (Optional, by default breakpoints are extracted using the AMR package, however if none are available or user wants to use a different one e.g. an ECOFF then it can be specified here).}

\item{use_mic}{Logical indicating whether to use MIC data in rule generation. Default is \code{TRUE}.}

\item{use_disk}{Logical indicating whether to use disk diffusion data for rule generation. Default is \code{TRUE}.}

\item{guide}{The guideline to use for breakpoints, when retrieving these from the AMR package. Default is \code{"EUCAST 2025"}.}

\item{makeRules}{Logical indicating whether to use define AMR rules from the data provided. Default is \code{TRUE}.}

\item{testRules}{Logical indicating whether to use test the rules by using them to interpret the genotypes as S/I/R and compare to the observed phenotypes. Default is \code{TRUE}.}

\item{expected_R}{Logical indicating whether to manually enforce whether or not this drug is an expected 'R' for this species. If NULL, we will check if this species & drug is included in \code{AMR::intrinsic_resistant}.}

\item{expected_I}{Logical indicating whether to manually enforce whether or not this drug should be reported as 'I' for this species. If NULL, we will check the S breakpoints and assume if this is set to 0.001 (MIC), or 50 (disk) this drug should always be reported as I not S.}
}
\value{
A data frame containing the generated AMR rules.
}
\description{
This function saves the results generated using \code{amrrules_analysis}, including figures and summary tables, to a specified directory.
It also uses these results to draft AMRrules using the \code{makerules} function and saves these to file
The output includes plots of reference distributions, PPV analysis, logistic regression, upset plots, and more, along with data summaries and rule tables.
}
\details{
This function saves the plots and data tables generated by \code{amrrules_analysis}, including:
\itemize{
\item Reference MIC and disk distribution plots (EUCAST)
\item PPV analysis plots
\item Logistic regression plots
\item Summary statistics tables
It then uses these data to draft AMRrules, returns the data and rules, and writes them to a TSV file for review.
}

The output is saved in a directory structure specified by \code{dir_path} and \code{outdir_name}.
If no output directory name is provided, the default is the antibiotic name.
}
\examples{
\dontrun{
amrrules_analysis(geno_table, pheno_table, antibiotic = "Ciprofloxacin",
                  drug_class_list = c("Quinolones"), species = "E. coli")
amrrules_save(amrrules, dir_path = "output/", outdir_name = "Ciprofloxacin")
}

}
