% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{categorise_from_PPV}
\alias{categorise_from_PPV}
\title{Classify markers as S/I/R and WT/NWT based on solo PPV values}
\usage{
categorise_from_PPV(data, suffix = "")
}
\arguments{
\item{data}{A data frame with column names including PPV estimates, named in the form of 'R.x.ppv', 'I.x.ppv', 'NWT.x.ppv'. This is intended as a helper function, called within the makerules() function.}

\item{suffix}{A character string indicating the suffix to append to the field names for the generated category (SIR) and phenotype (NWT/WT) fields.}
}
\value{
A copy of the input data frame with new fields indicating the SIR category and NWT/WT phenotype calls made from the PPV estimates.
}
\description{
This function takes as input a dataframe with columns indicating PPV values for R, I and/or NWT, and returns the data frame with new fields indicating the SIR and NWT/WT calls made from the PPV estimates.
}
\examples{
\dontrun{
data <- solo_stats \%>\%
 pivot_wider(id_cols=marker, names_from=category, values_from=c(ppv, x, n, ci.lower, ci.upper),
           names_glue= "{category}.solo.{.value}") \%>\%
 categorise_from_PPV(suffix="_soloPPV")
}

}
