% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{compare_categories}
\alias{compare_categories}
\title{Compare category calls and sample sizes for different sources of PPV (solo MIC+disk, extended solo, MIC, disk, logistic regression) and adjudicate a final call, recording evidence grade, limitations and notes for AMRrules. Helper function used by makerules().}
\usage{
compare_categories(
  category_soloPPV,
  category_soloExtPPV,
  category_micPPV,
  category_diskPPV,
  category_logReg,
  category_logRegExt,
  R.solo.n,
  R.soloExt.n,
  MIC.n,
  Disk.n,
  minObs = 1,
  low_threshold = 20
)
}
\arguments{
\item{category_soloPPV}{Call from solo PPV based on combined interpretation of MIC+disk assay data.}

\item{category_soloExtPPV}{Call from solo PPV based on all available SIR calls including those without raw assay data.}

\item{category_micPPV}{Call from solo PPV based on MIC data.}

\item{category_diskPPV}{Call from solo PPV based on disk data.}

\item{category_logReg}{Call from multivariable logistic regression including all markers exceeding a minimum frequency, for samples with assay measures.}

\item{category_logRegExt}{Call from multivariable logistic regression including all markers exceeding a minimum frequency, for all samples with SIR calls including those without raw assay data.}

\item{R.solo.n}{Number of samples contributing to category_soloPPV call.}

\item{R.soloExt.n}{Number of samples contributing to category_soloExtPPV call.}

\item{MIC.n}{Number of samples contributing to category_micPPV call.}

\item{Disk.n}{Number of samples contributing to category_diskPPV call.}

\item{minObs}{Minimum number of samples to make a call.}

\item{low_threshold}{Minimum number of observations required for \verb{evidence grade} to be assigned as "moderate" rather than "low". Default is 20.}
}
\value{
A list of values
}
\description{
Compare category calls and sample sizes for different sources of PPV (solo MIC+disk, extended solo, MIC, disk, logistic regression) and adjudicate a final call, recording evidence grade, limitations and notes for AMRrules. Helper function used by makerules().
}
\examples{
\dontrun{
data <- data \%>\%
  rowwise() \%>\%
  mutate(results=list(compare_categories(category_soloPPV=category_soloPPV,
             category_soloExtPPV=category_soloExtPPV,
             category_micPPV=category_micPPV,
             category_diskPPV=category_diskPPV,
             category_logReg=category_logReg,
             category_logRegExt=category_logRegExt,
             R.solo.n=R.solo.n, R.soloExt.n=R.soloExt.n,
             MIC.n=MIC.n, Disk.n=Disk.n))) \%>\%
 ungroup() \%>\%
 unnest_wider(results)
}

}
