% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeRules.R
\name{makerules}
\alias{makerules}
\title{Create AMR Rules Based on Geno-Pheno Data}
\usage{
makerules(
  amrrules,
  minObs = 3,
  low_threshold = 20,
  core_threshold = 0.9,
  use_mic = TRUE,
  mic_S = NULL,
  mic_R = NULL,
  mic_ecoff = NULL,
  use_disk = TRUE,
  disk_S = NULL,
  disk_R = NULL,
  disk_ecoff = NULL,
  guide = "EUCAST 2025",
  bp_site = NULL,
  rule_prefix = NULL,
  ruleID_start = 1000,
  note_prefix = "Quantitative geno-pheno analysis by ESGEM-AMR WG",
  regression = TRUE
)
}
\arguments{
\item{amrrules}{A list containing AMR analysis results, generated by \code{amrrules_analysis}.}

\item{minObs}{Minimum number of observations (samples) required for a given marker/combination, in order for PPV or assay data to be considered in rule setting. Default is 3.}

\item{low_threshold}{Minimum number of observations required for \verb{evidence grade} to be assigned as "moderate" rather than "low". Default is 20.}

\item{core_threshold}{The threshold used to classify a gene as "core" based on its frequency in the sample set. Default is 0.9 (90\%).}

\item{use_mic}{Logical indicating whether to use MIC data in rule generation. Default is \code{TRUE}.}

\item{mic_S}{The MIC S breakpoint. If \code{NULL}, breakpoints will be retrieved using the AMR package.}

\item{mic_R}{The MIC R breakpoint. If \code{NULL}, breakpoints will be retrieved using the AMR package}

\item{mic_ecoff}{The MIC ECOFF for the antibiotic, used to define nonwildtype. If \code{NULL}, ECOFF will be retrieved using the AMR package.}

\item{use_disk}{Logical indicating whether to use disk diffusion data for rule generation. Default is \code{TRUE}.}

\item{disk_S}{The disk diffusion S breakpoint. If \code{NULL}, breakpoints will be retrieved using the AMR package.}

\item{disk_R}{The disk diffusion R breakpoint. If \code{NULL}, breakpoints will be retrieved using the AMR package.}

\item{disk_ecoff}{The disk diffusion ECOFF for the antibiotic, used to define nonwildtype. If \code{NULL}, ECOFF will be retrieved using the AMR package.}

\item{guide}{The guideline to use for breakpoints, when retrieving these from the AMR package. Default is \code{"EUCAST 2025"}.}

\item{bp_site}{Site of infection, used if multiple breakpoint values are available for different body sites.}

\item{rule_prefix}{Prefix for the generated rule IDs. If \code{NULL}, prefix will be assigned based on the species name.}

\item{ruleID_start}{The starting ID number for generated rules. Default is 1000.}

\item{note_prefix}{Prefix for the note describing the method used to create the rules. Default is \code{"Quantitative geno-pheno analysis by ESGEM-AMR WG"}.}

\item{regression}{Logical indicating whether to consider logistic regression data (only used to define wildtype S for an individual marker, where the p-value exceeds 0.05 and odds ratio crosses 0, and there is no solo-marker data available). Default is \code{TRUE}.}
}
\value{
A list containing 2 data frames:
\itemize{
\item \code{rules}: The draft rules in AMRrules format (spec v0.6), with some additional quantitative data fields to support review.
\item \code{data}: Full set of quantitative data fields used to draft the rules.
}
}
\description{
This function generates AMRrules based on genotype-phenotype data.
The rules are created initially for single markers using solo PPV and median MIC/disk assay data in comparison to
the defined breakpoints and ECOFFs (either user-specified, or retrieved using the AMR package), and then
rules for combinations of markers are added if the phenotypes associated with the combination are more resistant
(R>I>S, nonwildtype>wildtype) than for the individual component markers.
}
\details{
The \code{makerules} function generates AMR rules based on quantitative analysis of genotype vs phenotype data.
The rules can be based on MIC or disk diffusion assay results, and different thresholds are applied to assign
evidence grades and limitations based on the quantiative data available, according to the AMRrules v0.5 specification.
A detailed \verb{rule curation note} is included in the output \code{rules} dataframe, summarising the quantitative
data supporting each rule.
Full data on each marker/combination is also returned in the separate \code{data} dataframe.
}
\examples{
\dontrun{
# Example of creating AMR rules
amrrules <- amrrules_analysis(geno_table, pheno_table, antibiotic = "Ciprofloxacin",
                              drug_class_list = c("Quinolones"), species = "E. coli")
rules <- makerules(amrrules)
}

}
\seealso{
\code{\link{amrrules_analysis}} for generating the AMR analysis data that is input into \code{makerules}.
}
