% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organism_codes.R
\docType{data}
\name{organism_codes}
\alias{organism_codes}
\title{Organism Code Reference Table}
\format{
A tibble with 45 rows and 3 columns:
\describe{
\item{Species}{The full species name, e.g. \code{"E. coli"}}
\item{Prefix}{A 3-letter code representing the organism, e.g. \code{"ECO"}}
\item{mo}{Microorganism ID as returned by \code{AMR::as.mo()}}
}
}
\source{
Manually curated for this package.
}
\usage{
data(organism_codes)
}
\description{
A dataset linking bacterial species names and AMR microorganisms (mo) to AMRrules
organism codes to use as prefixes for rules.
}
\examples{
data(organism_codes)
head(organism_codes)
dplyr::filter(organism_codes, Prefix == "ECO")

}
\keyword{datasets}
