% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concordance.R
\name{rules_concordance}
\alias{rules_concordance}
\title{Calculate genotype-phenotype concordance from rules-based predictions}
\usage{
rules_concordance(df, pheno_col = "pheno_eucast", pred_col = "category")
}
\arguments{
\item{pheno_col}{String indicating the name of the column containing the true phenotypes (default \code{"pheno_eucast"}).}

\item{pred_col}{String indicating the name of the column containing the predicted phenotypes (default \code{"category"}).}

\item{data}{A data frame containing observed and predicted phenotypes.}
}
\value{
A list containing:
\itemize{
\item \code{metrics}: A tibble with columns \code{metric}, \code{estimate}.
\item \code{matrix}: Confusion matrix comparing observed vs predicted.
}
}
\description{
Predictions are
compared to the observed phenotypes using standard classification metrics
(via the \code{yardstick} pkg) and AMR-specific error rates (major error, ME
and very major error, VME) per ISO 20776-2 (and see
\href{https://www.fda.gov/medical-devices/guidance-documents-medical-devices-and-radiation-emitting-products/antimicrobial-susceptibility-test-ast-systems-class-ii-special-controls-guidance-industry-and-fda}{FDA definitions}).
}
\details{
...

Standard metrics (sensitivity, specificity, PPV, NPV, accuracy, kappa,
F-measure) are calculated using pkg \code{yardstick}. AMR-specific error rates are
computed internally:
\itemize{
\item \strong{VME} (Very Major Error): FN / (TP + FN) = 1 - sensitivity. Proportion of
truly resistant isolates not predicted as such from genotype.
\item \strong{ME} (Major Error): FP / (TN + FP) = 1 - specificity. Proportion of
truly susceptible isolates incorrectly predicted resistant from genotype.
}
}
\examples{
\dontrun{
metrics_SIR <- rules_concordance(true_vs_predict))
metrics_NWT <- rules_concordance(true_vs_predict, "ecoff", "phenotype"))
}
}
\seealso{
\link[yardstick:yardstick-package]{yardstick::yardstick}
}
