% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{summarise_data}
\alias{summarise_data}
\title{Summarise Genotypic and Phenotypic Data for a given Antibiotic}
\usage{
summarise_data(
  geno_table,
  pheno_table,
  antibiotic,
  drug_class_list,
  geno_sample_col = "Name",
  pheno_sample_col = "id",
  species,
  guide = "EUCAST 2025"
)
}
\arguments{
\item{geno_table}{A data frame of genotypic data, including a column \code{drug_class} and sample IDs. Can be generated from AMRfinderplus output using \code{AMRgen::import_amrfp}}

\item{pheno_table}{A data frame of phenotypic data, including susceptibility results and sample IDs. Required fields are \code{drug_agent}, \code{mic}, \code{disk}.}

\item{antibiotic}{A string naming the antibiotic of interest (eg: \code{"Ciprofloxacin"}), will be parsed using AMR::as.ab().}

\item{drug_class_list}{A character vector of drug classes to include from the genotypic data (eg: \code{c("Quinolones")}).}

\item{geno_sample_col}{The name of the column in \code{geno_table} that contains sample IDs (default: "Name").}

\item{pheno_sample_col}{The name of the column in \code{pheno_table} that contains sample IDs (default: "id").}

\item{species}{The species for which EUCAST breakpoints should be retrieved (default: "E. coli"), will be parsed using AMR::as.mo().}

\item{guide}{The guidelines to retrieve breakpoints from using the AMR package (default: "EUCAST 2025").}
}
\value{
A tibble summarizing the number of samples with genotype and phenotype data, and listing EUCAST breakpoints and ECOFFs.
}
\description{
This function creates a summary of genotype and phenotype information for a specified antibiotic
and species using provided genotypic and phenotypic data tables. It reports sample counts with
relevant genetic markers and susceptibility test results, and lists EUCAST breakpoints and ECOFFs.
}
\details{
\itemize{
\item Uses \code{getBreakpoints()} to pull breakpoint and ECOFF data for the specified species, antibiotic, and guide.
\item Separately summarizes MIC and disk diffusion data if available.
}
}
\examples{
\dontrun{
summarise_data(geno_table, pheno_table, antibiotic = "Ciprofloxacin", drug_class_list=c("Quinolones"), species="E. coli")
}

}
